More code fixes and proper layouts
