+++ modules/messagedate.js
@@ -1,0 +1,40 @@
+var messageDate = function messageDate (message) {
+	var d = message.timestamp;
+	var hournow = d.getHours();
+	ampm = "AM";
+	if (hournow === 0) {
+		hournow = 12
+		ampm = "AM";
+	}
+	else if (hournow >= 13) {
+		hournow = hournow - 12;
+		ampm = "PM";
+	}
+	if (hournow < 10 && hournow > 0) {
+		hournow = "0" + hournow;
+	}
+	var minutenow = d.getMinutes();
+	if (minutenow < 10) {
+		minutenow = "0" + minutenow;
+	}
+	var secondnow = d.getSeconds();
+	if (secondnow < 10) {
+		secondnow = "0" + secondnow;
+	}
+	var day = d.getDate();
+	var monthIndex = d.getMonth();
+	var year = d.getFullYear();
+	var thedate = monthNames[monthIndex] + " " + day + ", " + year + " " + hournow + ":" + minutenow + ":" + secondnow + ampm;
+	return {
+		"thedate": thedate,
+		"year": year,
+		"month": monthNames[monthIndex],
+		"hour": hournow,
+		"minute": minutenow,
+		"ampm": ampm
+	};
+};
+
+module.exports = {
+  messageDate
+};
